refactor: move muiTheme into same file as component
