+++ src/features/Ontology/OntologyTabContainer.js
@@ -5,12 +5,34 @@ import Skeleton, { SkeletonTheme } from "react-loading-skeleton"
 import AppBar from "@material-ui/core/AppBar"
 import Tabs from "@material-ui/core/Tabs"
 import Tab from "@material-ui/core/Tab"
-import { MuiThemeProvider } from "@material-ui/core/styles"
-
+import { MuiThemeProvider, createMuiTheme } from "@material-ui/core/styles"
 import TypographyWrapper from "common/components/TypographyWrapper"
 import InnerGoPanel from "features/Ontology/InnerGoPanel"
 import { changeTab } from "./goaActions"
-import { goTabTheme as muiTheme } from "styles/goTabTheme"
+
+const muiTheme = createMuiTheme({
+  overrides: {
+    MuiTab: {
+      root: {
+        textTransform: "none",
+      },
+      // leaving this as an example of how to change selected tab properties
+      // selected: {
+      //   backgroundColor: "#f2f1ef",
+      // },
+    },
+    MuiTabs: {
+      root: {
+        backgroundColor: "#e6f2ff",
+        color: "#000",
+      },
+      indicator: {
+        backgroundColor: "#858780",
+        height: "3px",
+      },
+    },
+  },
+})
 
 type Props = {
   /** Object representing the "goa" slice of state */
